added tests/examples/
