+++ tests/allergy.test.js
@@ -2,7 +2,8 @@
 'use strict'
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
-
+const schema = require('./examples/allergy').schema
+const example = require('./examples/allergy').example
 const staticData = require('../filesObjects')
 // const allergies = require(allergiesFilePath)
 
@@ -11,16 +12,6 @@ describe('tests for allergy', () => {
     expect(staticData.allergies).not.toBe('')
   })
 })
-const schema = {
-  type: 'array', items: [{ type: 'string' }]
-}
-
-const example = [
-  'Gluten-Free',
-  'Peanut-Free',
-  'Seafood-Free',
-  'Sesame-Free'
-]
 
 describe('tests for allergy schema', () => {
   it('validates allergy json', () => {
