+++ tests/course.test.js
@@ -2,7 +2,8 @@
 'use strict'
 const { courseFilePath } = require('../files')
 const courses = require(courseFilePath)
-
+const schema = require('./examples/course').schema
+const example = require('./examples/course').example
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
@@ -11,26 +12,6 @@ describe('courses data files returns array', () => {
     expect(courses).not.toBe('')
   })
 })
-
-const schema = {
-  type: 'array', items: [{ type: 'string' }]
-}
-
-const example = [
-  'Main Dishes',
-  'Desserts',
-  'Side Dishes',
-  'Appetizers',
-  'Salads',
-  'Breakfast and Brunch',
-  'Breads',
-  'Soups',
-  'Beverages',
-  'Condiments and Sauces',
-  'Cocktails',
-  'Snacks',
-  'Lunch'
-]
 
 describe('tests for course schema', () => {
   it('validates course json', () => {
