+++ tests/cuisine.test.js
@@ -2,7 +2,8 @@
 'use strict'
 const { cuisineFilePath } = require('../files')
 const cuisines = require(cuisineFilePath)
-
+const schema = require('./examples/cuisine').schema
+const example = require('./examples/cuisine').example
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
@@ -11,39 +12,6 @@ describe('this test prevents to any issues and problems, also to break the struc
     expect(cuisines).not.toBe('')
   })
 })
-
-const schema = {
-  type: 'array', items: [{ type: 'string' }]
-}
-
-const example = [
-  'American',
-  'Kid-Friendly',
-  'Italian',
-  'Asian',
-  'Mexican',
-  'Southern & Soul Food',
-  'French',
-  'Southwestern',
-  'Barbecue',
-  'Indian',
-  'Chinese',
-  'Cajun & Creole',
-  'Mediterranean',
-  'Greek',
-  'English',
-  'Spanish',
-  'Thai',
-  'German',
-  'Moroccan',
-  'Irish',
-  'Japanese',
-  'Cuban',
-  'Hawaiian',
-  'Swedish',
-  'Hungarian',
-  'Portuguese'
-]
 
 describe('tests for cuisine schema', () => {
   it('validates cuisine json', () => {
