+++ tests/cup1.test.js
@@ -9,40 +9,14 @@ const {
   cup1
 } = require('../filesObjects')
 
+const schema = require('./examples/cup1').schema
+const example = require('./examples/cup1').example
+
 describe('this test prevents to any issues and problems, also to break the structure of cup1 data', () => {
   it('cup1 data files returns object', () => {
     expect(cup1).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: {
-      type: 'string'
-    },
-    americanStandart: {
-      type: 'string'
-    },
-    americanStandartOunces: {
-      type: 'string'
-    },
-    volume: {
-      type: 'string'
-    },
-    weight: {
-      type: 'string'
-    }
-  },
-  required: ['name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight']
-}
-
-const example = {
-  name: 'cup',
-  americanStandart: '16 tablespoons',
-  americanStandartOunces: '8 ounces',
-  volume: '240 milliliters',
-  weight: '225 grams,  1/2 pound'
-}
 
 describe('test cup1 json schema', () => {
   it('validates my jason', () => {
