+++ tests/cup1_8.test.js
@@ -9,40 +9,14 @@ const {
   cup18
 } = require('../filesObjects')
 
+const schema = require('./examples/cup1_8').schema
+const example = require('./examples/cup1_8').example
+
 describe('this test prevents to any issues and problems, also to break the structure of cup1_8 data', () => {
   it('cup1_8 data files returns object', () => {
     expect(cup18).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: {
-      type: 'string'
-    },
-    americanStandart: {
-      type: 'string'
-    },
-    americanStandartOunces: {
-      type: 'string'
-    },
-    volume: {
-      type: 'string'
-    },
-    weight: {
-      type: 'string'
-    }
-  },
-  required: ['name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight']
-}
-
-const example = {
-  name: 'cup1_8',
-  americanStandart: '2 tablespoons',
-  americanStandartOunces: '1 ounce',
-  volume: '30 milliliters',
-  weight: '28.3 grams'
-}
 
 describe('test cup1_8 json schema', () => {
   it('validates my json', () => {
