+++ tests/cup3_4.test.js
@@ -2,7 +2,8 @@
 'use strict'
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
-
+const schema = require('./examples/cup3_4').schema
+const example = require('./examples/cup3_4').example
 const { cup34 } = require('../filesObjects')
 
 describe('this test prevents to any issues and problems, also to break the structure of cup3_4 data', () => {
@@ -10,21 +11,6 @@ describe('this test prevents to any issues and problems, also to break the struc
     expect(cup34).not.toBe('')
   })
 })
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'string' },
-    americanStandartOunces: { type: 'string' },
-    volume: { type: 'string' },
-    weight: { type: 'string' }
-  },
-  required: ['name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight']
-}
-const example = { name: 'cup3_4',
-  americanStandart: '12 tablespoons',
-  americanStandartOunces: '6 ounces',
-  volume: '180 milliliters',
-  weight: 'null' }
 
 describe('test cup3_4 json schema', () => {
   it('validates my json', () => {
