+++ tests/examples/allergy.js
@@ -1,0 +1,14 @@
+const schema = {
+  type: 'array', items: [{ type: 'string' }]
+}
+
+const example = [
+  'Gluten-Free',
+  'Peanut-Free',
+  'Seafood-Free',
+  'Sesame-Free'
+]
+
+module.exports = {
+  schema, example
+}
