+++ tests/examples/course.js
@@ -1,0 +1,23 @@
+
+const schema = {
+  type: 'array', items: [{ type: 'string' }]
+}
+
+const example = [
+  'Main Dishes',
+  'Desserts',
+  'Side Dishes',
+  'Appetizers',
+  'Salads',
+  'Breakfast and Brunch',
+  'Breads',
+  'Soups',
+  'Beverages',
+  'Condiments and Sauces',
+  'Cocktails',
+  'Snacks',
+  'Lunch'
+]
+module.exports = {
+  example, schema
+}
