+++ tests/examples/cuisine.js
@@ -1,0 +1,37 @@
+
+const schema = {
+  type: 'array', items: [{ type: 'string' }]
+}
+
+const example = [
+  'American',
+  'Kid-Friendly',
+  'Italian',
+  'Asian',
+  'Mexican',
+  'Southern & Soul Food',
+  'French',
+  'Southwestern',
+  'Barbecue',
+  'Indian',
+  'Chinese',
+  'Cajun & Creole',
+  'Mediterranean',
+  'Greek',
+  'English',
+  'Spanish',
+  'Thai',
+  'German',
+  'Moroccan',
+  'Irish',
+  'Japanese',
+  'Cuban',
+  'Hawaiian',
+  'Swedish',
+  'Hungarian',
+  'Portuguese'
+]
+
+module.exports = {
+  schema, example
+}
