+++ tests/examples/department.js
@@ -1,0 +1,20 @@
+const schema = {
+  properties: {
+    name: {
+      type: 'string'
+    },
+    type: {
+      type: 'string'
+    }
+
+  },
+  required: ['name', 'type']
+}
+let example = {
+  'name': 'Fresh vegetables',
+  'type': 'food'
+}
+
+module.exports = {
+  schema, example
+}
