+++ tests/favorites.test.js
@@ -6,33 +6,15 @@ const favorites = require(favoritesFilePath)
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
+const schema = require('./examples/favorites').schema
+const example = require('./examples/favorites').example
+
 describe('favorites data files returns array', () => {
   it('these tests prevent any issues and problems, also to break the structure of favorites', () => {
     expect(favorites).not.toBe('')
   })
 })
-const schema = {
-  properties: {
-    ingredient_id: {
-      type: 'string'
-    },
-    user_id: {
-      type: 'string'
-    },
-    favs: {
-      type: 'string'
-    },
-    grocery_id: {
-      type: 'string'
-    }
-  },
-  required: ['ingredient_id', 'user_id', 'favs', 'grocery_id']
-}
-const example = { 'ingredient_id': '4aacafd5-6106-11e9-9ffc-a3e52b2d927f',
-  'user_id': '4aacafd2-6106-11e9-9ffc-a3e52b2d927f',
-  'favs': 'desc for department1',
-  'grocery_id': '4aac88cf-6106-11e9-9ffc-a3e52b2d927f'
-}
+
 describe('tests favorites json schema', () => {
   it('validates my json', () => {
     expect(example).toMatchSchema(schema)
